+++ src/simp-algorithms/TRACE_Q.hpp
@@ -73,8 +73,8 @@ namespace trace_q {
             double x_high{};
             double y_low{};
             double y_high{};
-            long double t_low{};
-            long double t_high{};
+            unsigned long t_low{};
+            unsigned long t_high{};
         };
 
         /**
